LCD and Game Creation
The LCD has been created with the game. As of now, the Transmitting HM-10 Bluetooth Module is sending a random 1 or 0 every second, 1 representing a flex. The spikes move from left to right and the character can move between the two lanes to dodge the spikes. Collision detection has not been added either.